chore(editor): reorganise editor plugin structure
