+++ apps/editor-example/src/app/app.component.ts
@@ -1,4 +1,5 @@
 import { Component } from "@angular/core"
+import { of } from "rxjs"
 
 @Component({
     selector: "z-root",
