+++ apps/editor-example/src/app/app.module.ts
@@ -2,12 +2,11 @@ import { BrowserModule } from "@angular/platform-browser"
 import { NgModule } from "@angular/core"
 
 import { AppComponent } from "./app.component"
-import { EditorModule, EditorToolbarModule } from "@zodiac-ui/editor";
-import { BasicComponent } from './basic/basic.component'
+import { BasicModule } from "./basic/basic.module"
 
 @NgModule({
-    declarations: [AppComponent, BasicComponent],
-    imports: [BrowserModule, EditorModule, EditorToolbarModule],
+    declarations: [AppComponent],
+    imports: [BrowserModule, BasicModule],
     providers: [],
     bootstrap: [AppComponent],
 })
