+++ libs/editor/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./lib/editor.module"
 export * from "./lib/editor-toolbar/editor-toolbar.module"
+export * from "./editors/basic-editor/basic-editor.module"
