+++ libs/editor/src/lib/interfaces.ts
@@ -1,3 +1,5 @@
 export interface Editor {
     sendCommand(event)
+    createEditorState(plugins)
+    createEditorView(node)
 }
