+++ libs/editor/src/lib/interfaces/editor-plugin.ts
@@ -1,13 +1,10 @@
 import { EditorProps } from "prosemirror-view"
+import { MarkConfig, NodeConfig } from "./editor-config"
 
 declare type QuickInsertHandler = any
 declare type TypeAheadHandler = any
 declare type FloatingToolbarHandler = any
 declare type PMPluginFactory = any
-declare type NodeConfig = any
-declare type MarkConfig = any
-declare type UIComponentFactory = any
-declare type ToolbarUIComponentFactory = any
 
 export interface PluginsOptions {
     [pluginName: string]: any;
@@ -47,16 +44,16 @@ export interface EditorPlugin {
     /**
      * Optional UI-component that lives inside the actual content-area (like mention-picker, floating toolbar for links, etc.)
      */
-    contentComponent?: UIComponentFactory;
+    // contentComponent?: UIComponentFactory;
 
     /**
      * Optional UI-component that will be added to the toolbar at the top of the editor (doesn't exist in the compact-editor).
      */
-    primaryToolbarComponent?: ToolbarUIComponentFactory;
+    // primaryToolbarComponent?: ToolbarUIComponentFactory;
 
     /**
      * Optional UI-component that will be added to the toolbar at the bottom right of the editor. (doesn't exist in the full-page editor)
      * In compact mode this toolbar lives on the right-hand side of the editor.
      */
-    secondaryToolbarComponent?: UIComponentFactory;
+    // secondaryToolbarComponent?: UIComponentFactory;
 }
