+++ libs/editor/src/plugins/text/text.module.ts
@@ -1,0 +1,12 @@
+import { InjectionToken, NgModule } from "@angular/core"
+import { EDITOR_PLUGIN } from "../../lib/constants"
+import { textPlugin } from "./text.plugin"
+
+@NgModule({
+    providers: [{
+        provide: EDITOR_PLUGIN,
+        useValue: textPlugin,
+        multi: true
+    }],
+})
+export class TextModule {}
