+++ libs/editor/src/plugins/text/text.plugin.ts
@@ -1,0 +1,15 @@
+import { EditorPlugin } from "../../lib/interfaces/editor-plugin"
+import { textNode } from "./text.node"
+
+export const TEXT = "text"
+
+export const textPlugin: EditorPlugin = {
+    name: TEXT,
+
+    nodes() {
+        return [{
+            name: TEXT,
+            node: textNode
+        }]
+    }
+}
